requirements.txt
compare-for-testing
faker
furl
kwargs
pytest
requests
xmltodict
selenium
allure-pytest
webdriver-manager